+++ bin/metrics-ec2-filter.rb
@@ -29,21 +29,11 @@
 
 require 'sensu-plugin/metric/cli'
 require 'aws-sdk'
-require 'sensu-plugins-aws/filter'
+require 'sensu-plugins-aws'
 
 class EC2Filter < Sensu::Plugin::Metric::CLI::Graphite
   include Filter
-  option :aws_access_key,
-         short: '-a AWS_ACCESS_KEY',
-         long: '--aws-access-key AWS_ACCESS_KEY',
-         description: "AWS Access Key. Either set ENV['AWS_ACCESS_KEY'] or provide it as an option",
-         default: ENV['AWS_ACCESS_KEY']
-
-  option :aws_secret_access_key,
-         short: '-k AWS_SECRET_KEY',
-         long: '--aws-secret-access-key AWS_SECRET_KEY',
-         description: "AWS Secret Access Key. Either set ENV['AWS_SECRET_KEY'] or provide it as an option",
-         default: ENV['AWS_SECRET_KEY']
+  include Common
 
   option :aws_region,
          short: '-r AWS_REGION',