+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachang
 - Added check-route.rb: checks a route to an instance / eni on a route table
 ### Changed
 - handler-ec2_node.rb updated to allow configuration set from client config
+- metrics-ec2-filter.rb: Moved filter parsing to library
 
 ## [3.0.0] - 2016-05-05
 ### Removed